refactor: moved column sorting logic to its own module
