+++ packages/rxtree/src/operators/cartTree.js
@@ -1,30 +1,5 @@
-import sortBy from 'lodash/sortBy';
-import {zip} from 'rxjs';
-import {map,mergeMap,scan,takeLast,tap} from 'rxjs/operators';
-
 import giniGain from './giniGain';
-
-function createSortedColumnsFromRows() {
-  return row$ => row$.pipe(
-    scan(([index], [features, label]) => [
-      index + 1,
-      features,
-      label,
-    ], [-1, null, null]),
-    map(([index, features, label]) => (
-      features.map(feature => ({index, feature, label}))
-    )),
-    tap(console.log),
-    scan((dataframe, columns) => {
-      if (!dataframe) return columns.map(c => [c]);
-      columns.map((colVal, colIndex) => dataframe[colIndex].push(colVal));
-      return dataframe;
-    }, null),
-    tap(console.log),
-    takeLast(1),
-    map(df => df.map(column => sortBy(column, val => val.feature)))
-  );
-}
+import createSortedColumnsFromRows from '../internals/createSortedColumnsFromRows';
 
 const cartTree = function cartTree({
   initialState = {},
